+++ packages/clerk-js/src/core/resources/AuthConfig.ts
@@ -4,6 +4,7 @@ import { BaseResource } from './internal';
 
 export class AuthConfig extends BaseResource implements AuthConfigResource {
   singleSessionMode!: boolean;
+  cookielessDev!: boolean;
 
   public constructor(data: AuthConfigJSON) {
     super();
@@ -12,6 +13,7 @@ export class AuthConfig extends BaseResource implements AuthConfigResource {
 
   protected fromJSON(data: AuthConfigJSON | null): this {
     this.singleSessionMode = data ? data.single_session_mode : true;
+    this.cookielessDev = data ? data.cookieless_dev : false;
     return this;
   }
 }