+++ packages/clerk-js/src/utils/index.ts
@@ -13,6 +13,7 @@ export * from './jwt';
 export * from './pageLifecycle';
 export * from './path';
 export * from './querystring';
+export * from './runtime';
 export * from './runWithExponentialBackOff';
 export * from './safeLock';
 export * from './script';