+++ packages/react/src/types.ts
@@ -16,6 +16,10 @@ export interface BrowserClerkConstructor {
   new (frontendApi: string): BrowserClerk;
 }
 
+export interface HeadlessBrowserClerkConstrutor {
+  new (frontendApi: string): HeadlessBrowserClerk;
+}
+
 export type WithClerkProp<T> = T & { clerk: LoadedClerk };
 
 export type WithUserProp<T> = T & { user: UserResource };
@@ -28,14 +32,23 @@ export interface MountProps {
   props?: any;
 }
 
-export interface BrowserClerk extends Clerk {
+export interface HeadlessBrowserClerk extends Clerk {
   load: (opts?: ClerkOptions) => Promise<void>;
   updateClient: (client: ClientResource) => void;
+}
+
+export interface BrowserClerk extends HeadlessBrowserClerk {
   onComponentsReady: Promise<void>;
   components: any;
 }
 
-export type ClerkProp = BrowserClerkConstructor | BrowserClerk | undefined | null;
+export type ClerkProp =
+  | BrowserClerkConstructor
+  | BrowserClerk
+  | HeadlessBrowserClerk
+  | HeadlessBrowserClerkConstrutor
+  | undefined
+  | null;
 
 type ButtonProps = {
   afterSignInUrl?: string;