+++ playground/nextjs/pages/api/hello.ts
@@ -1,10 +1,13 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import type { NextApiRequest, NextApiResponse } from 'next';
+import { getAuth } from '@clerk/nextjs/server';
 
 type Data = {
   name: string;
 };
 
 export default function handler(req: NextApiRequest, res: NextApiResponse<Data>) {
-  res.status(200).json({ name: 'John Doe' });
+  const { userId } = getAuth(req);
+  console.log('/api/hello ', userId, req.query);
+  res.status(200).json({ userId });
 }