+++ packages/sdk-node/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [4.5.3-staging.2](https://github.com/clerkinc/javascript/compare/@clerk/clerk-sdk-node@4.5.3-staging.1...@clerk/clerk-sdk-node@4.5.3-staging.2) (2022-11-21)
+
+**Note:** Version bump only for package @clerk/clerk-sdk-node
+
+### [4.5.2](https://github.com/clerkinc/javascript/compare/@clerk/clerk-sdk-node@4.5.2-staging.1...@clerk/clerk-sdk-node@4.5.2) (2022-11-18)
+
+**Note:** Version bump only for package @clerk/clerk-sdk-node
+
 ### [4.5.1](https://github.com/clerkinc/javascript/compare/@clerk/clerk-sdk-node@4.5.1-staging.3...@clerk/clerk-sdk-node@4.5.1) (2022-11-15)
 
 **Note:** Version bump only for package @clerk/clerk-sdk-node